requirements.txt
fastapi 
uvicorn 
langgraph 
langchain 
google-generativeai 
python-dotenv
google-genai
langchain-google-genai
streamlit
requests

groq